Refactored Everything into IntakeMap
We refactored the intake controller into intakemap and added more state
handling stuff.
